try3

+++ hmt_processor/processors.py
@@ -512,14 +512,14 @@ def hmt_tile_binary_processor_griddedHMT(tile_path, hmt_incriment_tile_path, out
       # Build job here
       #logger.info("    Working on block offset ({0},{1}); cols {2}; rows: {3}...".format(j,i, numCols, numRows))
       
-      lidar_np = tile_lidar.ReadAsArray(j, i, numCols, numRows)
-      hmt_np = tile_hmt.ReadAsArray(j, i, numCols, numRows)
+      lidar_np = tile_lidar.ReadAsArray(j, i, numCols, numRows).astype(np.float)
+      hmt_np = tile_hmt.ReadAsArray(j, i, numCols, numRows).astype(np.float)
       
       lidar_hmt_masked_below_hmt = np.ma.masked_equal(lidar_np, tile_nodata, copy=True).filled(np.nan)  # Create the mask
       binary_rast = lidar_hmt_masked_below_hmt <= hmt_np
       
       # Write the array to the raster
-      HMT_output_band.WriteArray(binary_rast, j, i)
+      HMT_output_band.WriteArray(binary_rast.astype(np.int), j, i)
       
       # Clean Up
       HMT_output_fh.FlushCache()
